ADR-0098 R7: drop redundant u64 cast in regression test (clippy -D warnings)
verify_chain() already returns u64; `verified as u64` tripped
clippy::unnecessary_cast under -D warnings. Build + Test steps passed on both
arms (the regression test compiled and passed) — this was the only lint.

+++ apps/aberp/tests/adr0098_r7_backfill_refork.rs
@@ -206,7 +206,7 @@ fn r7_backfill_cycle_via_shared_handle_is_coherent_no_refork() {
         .expect("reopen ledger for verify");
     let verified = ledger.verify_chain().expect("chain verifies");
     assert_eq!(
-        verified as u64, burst,
+        verified, burst,
         "verify_chain clean across the whole Handle-routed burst"
     );
